module_6/src/web/requirements.txt
anyio==4.12.1
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
certifi==2026.1.4
click==8.3.1
diskcache==5.6.3
filelock==3.20.3
Flask==3.1.2
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
packaging==26.0
pika==1.3.2
psycopg[binary]==3.3.2
PyYAML==6.0.3
python-dotenv==1.2.1
shellingham==1.5.4
soupsieve==2.8.3
tqdm==4.67.2
typer-slim==0.21.1
typing_extensions==4.15.0
Werkzeug==3.1.5
